Added rendering environments/Elaborate scene structure
- Added SSAO code, commented out for now
- Render environments to 'hardcode' techniques, but provide an
abstraction nonetheless
