+++ OpenGLRendererProject/OpenGLRendererProject/ComputeRaytracerEnvironment.cpp
@@ -1,0 +1,27 @@
+#include "ComputeRaytracerEnvironment.h"
+
+using namespace Primitive;
+
+ComputeRaytracerEnvironment::ComputeRaytracerEnvironment()
+{
+}
+
+
+ComputeRaytracerEnvironment::~ComputeRaytracerEnvironment()
+{
+}
+
+void ComputeRaytracerEnvironment::Initialize()
+{
+
+}
+
+void ComputeRaytracerEnvironment::Render(IScene* aScene)
+{
+
+}
+
+void ComputeRaytracerEnvironment::CleanUp()
+{
+
+}
