+++ OpenGLRendererProject/OpenGLRendererProject/ComputeRaytracerEnvironment.h
@@ -1,0 +1,20 @@
+#pragma once
+#include "IRenderEnvironment.h"
+
+namespace Primitive
+{
+
+
+class ComputeRaytracerEnvironment :
+	public IRenderEnvironment
+{
+public:
+	ComputeRaytracerEnvironment();
+	~ComputeRaytracerEnvironment();
+
+	void Initialize();
+	void Render(IScene* aScene);
+	void CleanUp();
+};
+
+};
