+++ OpenGLRendererProject/OpenGLRendererProject/IRenderEnvironment.cpp
@@ -1,0 +1,12 @@
+#include "IRenderEnvironment.h"
+
+using namespace Primitive;
+
+IRenderEnvironment::IRenderEnvironment()
+{
+}
+
+
+IRenderEnvironment::~IRenderEnvironment()
+{
+}
