+++ OpenGLRendererProject/OpenGLRendererProject/IScene.h
@@ -6,6 +6,9 @@
 #include "Light.h"
 #include "Input.h"
 #include "Camera.h"
+#include "SceneObject.h"
+#include "ModelLoader.h"
+#include "GLTextureLoader.h"
 
 namespace Primitive{
 
@@ -21,6 +24,7 @@ public:
 	virtual ~IScene() = default;
 
 	// Get deltatime and input every frame
+	virtual void UploadModels(ModelLoader* const aModelLoader, GLTextureLoader* const aTextureLoader) = 0;
 	virtual void Update(float aDT, std::shared_ptr<Input> aInput) = 0;
 	virtual void OnMouseMove(std::shared_ptr<Input> aInput) = 0;
 
@@ -28,13 +32,19 @@ public:
 
 	const std::vector<Light>& GetAllLights();
 	const std::shared_ptr<Camera> GetCamera();
+	const std::vector<std::shared_ptr<SceneObject>> GetSceneObjects();
+
 	// Maybe protected destroy? To clear scenes in the destructor
 	// void Destroy();
+
 protected:
+	// Just point lights for now
 	std::vector<Light> mLights;
 
 	// Only support one camera for now
 	std::shared_ptr<Camera> mCamera;
+
+	std::vector<std::shared_ptr<SceneObject>> mSceneObjects;
 };
 
 }
