+++ OpenGLRendererProject/OpenGLRendererProject/OpenGLRendererProject.vcxproj
@@ -162,7 +162,9 @@
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="AssetsDatabase.h" />
+    <ClInclude Include="BloomEnvironment.h" />
     <ClInclude Include="Camera.h" />
+    <ClInclude Include="ComputeRaytracerEnvironment.h" />
     <ClInclude Include="EngineTimer.h" />
     <ClInclude Include="FirstTestScene.h" />
     <ClInclude Include="GLFramebuffer.h" />
@@ -171,18 +173,24 @@
     <ClInclude Include="GLTextureLoader.h" />
     <ClInclude Include="GLTimer.h" />
     <ClInclude Include="Input.h" />
+    <ClInclude Include="IRenderEnvironment.h" />
     <ClInclude Include="Light.h" />
     <ClInclude Include="MeshData.h" />
     <ClInclude Include="ModelLoader.h" />
     <ClInclude Include="IScene.h" />
     <ClInclude Include="Renderer.h" />
+    <ClInclude Include="RenderingHelperFunctions.h" />
+    <ClInclude Include="SceneObject.h" />
     <ClInclude Include="Shader.h" />
     <ClInclude Include="ShaderProgram.h" />
     <ClInclude Include="stdafx.h" />
     <ClInclude Include="targetver.h" />
+    <ClInclude Include="TextureSamplers.h" />
   </ItemGroup>
   <ItemGroup>
+    <ClCompile Include="BloomEnvironment.cpp" />
     <ClCompile Include="Camera.cpp" />
+    <ClCompile Include="ComputeRaytracerEnvironment.cpp" />
     <ClCompile Include="EngineTimer.cpp" />
     <ClCompile Include="FirstTestScene.cpp" />
     <ClCompile Include="GLFramebuffer.cpp" />
@@ -191,10 +199,12 @@
     <ClCompile Include="GLTextureLoader.cpp" />
     <ClCompile Include="GLTimer.cpp" />
     <ClCompile Include="Input.cpp" />
+    <ClCompile Include="IRenderEnvironment.cpp" />
     <ClCompile Include="IScene.cpp" />
     <ClCompile Include="ModelLoader.cpp" />
     <ClCompile Include="OpenGLRendererProject.cpp" />
     <ClCompile Include="Renderer.cpp" />
+    <ClCompile Include="SceneObject.cpp" />
     <ClCompile Include="Shader.cpp" />
     <ClCompile Include="ShaderProgram.cpp" />
     <ClCompile Include="stdafx.cpp" />
@@ -203,10 +213,14 @@
     <None Include="Shaders\FragmentShaderBasicPosition.frag" />
     <None Include="Shaders\FragmentShaderLight.frag" />
     <None Include="Shaders\GaussianBlur.vert" />
+    <None Include="Shaders\GBufferViewPass.frag" />
+    <None Include="Shaders\GBufferViewPass.vert" />
     <None Include="Shaders\GuassianBlur.frag" />
     <None Include="Shaders\InitialTestCompute.comp" />
     <None Include="Shaders\PostProcHDR.frag" />
     <None Include="Shaders\PostProcHDR.vert" />
+    <None Include="Shaders\ShaderSSAO.frag" />
+    <None Include="Shaders\ShaderSSAO.vert" />
     <None Include="Shaders\VertexShaderBasicPosition.vert" />
     <None Include="Shaders\VertexShaderLight.vert" />
   </ItemGroup>
