+++ OpenGLRendererProject/OpenGLRendererProject/SceneObject.cpp
@@ -1,0 +1,14 @@
+#include "SceneObject.h"
+
+#include "GLMesh.h"
+
+using namespace Primitive;
+
+SceneObject::SceneObject() : mTransform(glm::mat4())
+{
+}
+
+
+SceneObject::~SceneObject()
+{
+}
