+++ OpenGLRendererProject/OpenGLRendererProject/stdafx.h
@@ -17,6 +17,7 @@
 #include <assert.h>
 #include <inttypes.h>
 #include <string>
+#include <random>
 
 
 // TODO: reference additional headers your program requires here
